Move blend_context.rs pub methods to blend.rs,
Functions, get_values, get_value, get_table_values are only used in blend.rs.
So those should be in blend.rs and also private
